Test: 피드 테스트를 단위테스트와 통합테스트로 분할
