+++ src/test/java/com/stoury/service/FeedServiceTest.java
@@ -49,14 +49,6 @@ public class FeedServiceTest {
     FeedService feedService;
     Member writer;
 
-    @BeforeEach
-    @AfterEach
-    void tearDown() {
-        feedRepository.deleteAll();
-        memberRepository.deleteAll();
-        tagRepository.deleteAll();
-    }
-
     @Test
     @DisplayName("피드 생성 성공")
     void createFeedSuccess() {
@@ -66,12 +58,11 @@ public class FeedServiceTest {
                 .longitude(111.111)
                 .latitude(333.333)
                 .build();
-        Member member = Member.builder()
+        writer = Member.builder()
                 .username("writer")
                 .encryptedPassword("dqwdasda")
                 .email("d1wd@wdwfef.com")
                 .build();
-        Member writer = memberRepository.save(member);
 
         List<MultipartFile> graphicContents = List.of(
                 new MockMultipartFile("Files", "first", "image/jpeg", new byte[0]),
@@ -82,6 +73,7 @@ public class FeedServiceTest {
                 .textContent(feedCreateRequest.textContent())
                 .longitude(feedCreateRequest.longitude())
                 .latitude(feedCreateRequest.latitude())
+                .tags(new ArrayList<>())
                 .build());
 
         FeedResponse createdFeed = feedService.createFeed(writer, feedCreateRequest, graphicContents);
@@ -95,76 +87,14 @@ public class FeedServiceTest {
     }
 
     @Test
-    @DisplayName("피드 생성과 동시에 태그 생성 성공")
-    void createFeedWithTagsSuccess() {
-        // 태그가 없는 상태에서 시작
-        assertThat(tagRepository.findAll()).isEmpty();
-
-        Member member = Member.builder()
-                .username("writer")
-                .encryptedPassword("dqwdasda")
-                .email("d1wd@wdwfef.com")
-                .build();
-        Member writer = memberRepository.save(member);
-
-        List<String> tagGroup = List.of("tag1", "tag2", "tag3");
-        FeedCreateRequest feedCreateRequest = FeedCreateRequest.builder()
-                .textContent("with tag group1")
-                .longitude(11.11)
-                .latitude(11.11)
-                .tagNames(tagGroup)
-                .build();
-        List<MultipartFile> graphicContents = List.of(
-                new MockMultipartFile("Files", "first", "image/jpeg", new byte[0]),
-                new MockMultipartFile("Files", "second", "video/mp4", new byte[0])
-        );
-
-        FeedResponse feedResponse = feedService.createFeed(writer, feedCreateRequest, graphicContents);
-
-        assertThat(feedResponse.tagNames()).containsExactly(tagGroup.toArray(String[]::new));
-    }
-
-    @Test
-    @DisplayName("이미 존재하는 태그로 피드 생성 성공")
-    void createFeedWithExistTags() {
-        Member member = Member.builder()
-                .username("writer")
-                .encryptedPassword("dqwdasda")
-                .email("d1wd@wdwfef.com")
-                .build();
-        Member writer = memberRepository.save(member);
-
-        List<String> tagGroup = tagRepository.saveAll(List.of(
-                new Tag("tag1"),
-                new Tag("tag2"),
-                new Tag("tag3"))).stream().map(Tag::getTagName).toList();
-
-        FeedCreateRequest feedCreateRequest = FeedCreateRequest.builder()
-                .textContent("with tag group1")
-                .longitude(11.11)
-                .latitude(11.11)
-                .tagNames(tagGroup)
-                .build();
-        List<MultipartFile> graphicContents = List.of(
-                new MockMultipartFile("Files", "first", "image/jpeg", new byte[0]),
-                new MockMultipartFile("Files", "second", "video/mp4", new byte[0])
-        );
-
-        FeedResponse feedResponse = feedService.createFeed(writer, feedCreateRequest, graphicContents);
-
-        assertThat(feedResponse.tagNames()).containsExactly(tagGroup.toArray(String[]::new));
-    }
-
-    @Test
     @DisplayName("피드 생성 실패, 지원하지 않는 파일")
     @Transactional
     void createFeedFailByNotSupportedFileFormat() {
-        Member member = Member.builder()
+        writer = Member.builder()
                 .username("writer")
                 .encryptedPassword("dqwdasda")
                 .email("d1wd@wdwfef.com")
                 .build();
-        Member writer = memberRepository.save(member);
 
         FeedCreateRequest createFeed = FeedCreateRequest.builder()
                 .textContent("testing")
@@ -185,12 +115,11 @@ public class FeedServiceTest {
     @DisplayName("피드 저장 실패, 이미지 없음")
     @Transactional
     void createFeedFailNoGraphicContents() {
-        Member member = Member.builder()
+        writer = Member.builder()
                 .username("writer")
                 .encryptedPassword("dqwdasda")
                 .email("d1wd@wdwfef.com")
                 .build();
-        Member writer = memberRepository.save(member);
 
         FeedCreateRequest createFeed = FeedCreateRequest.builder()
                 .textContent("testing")
@@ -212,79 +141,5 @@ public class FeedServiceTest {
                 .build()).isInstanceOf(FeedCreateException.class);
     }
 
-    @Test
-    @DisplayName("태그로 피드 조회")
-    @Transactional
-    void searchFeedsByTag() {
-        String tagName = "testTag";
-        Tag tag = new Tag(tagName);
-        tagRepository.save(tag);
-        List<Feed> feeds = create20FeedsWith(tag);
-        feedRepository.saveAll(feeds);
 
-        LocalDateTime maxDateTime = LocalDate.of(2100, 12, 31).atStartOfDay();
-        Slice<FeedResponse> recent10Feeds = feedService.getFeedsByTag(tagName, maxDateTime);
-
-        assertThat(recent10Feeds).hasSize(10);
-        assertThat(recent10Feeds.hasNext()).isTrue();
-
-        List<FeedResponse> recentFeedResponses = recent10Feeds.getContent();
-        for (int i = 0; i < recentFeedResponses.size(); i++) {
-            FeedResponse feedResponse = recentFeedResponses.get(i);
-            assertThat(feedResponse.textContent()).isEqualTo(feeds.get(feeds.size() - i - 1).getTextContent());
-        }
-    }
-
-    @Test
-    @DisplayName("사용자 피드 조회")
-    void searchFeedsOfMember() {
-        Member member = Member.builder()
-                .username("writer")
-                .encryptedPassword("dqwdasda")
-                .email("d1wd@wdwfef.com")
-                .build();
-        Member writer = memberRepository.save(member);
-
-        List<Feed> feeds = IntStream.range(0, 20).mapToObj(i -> Feed.builder()
-                .textContent("feed" + i)
-                .latitude(11.11)
-                .longitude(11.11)
-                .member(writer)
-                .build()).toList();
-        feedRepository.saveAll(feeds);
-
-        LocalDateTime maxDateTime = LocalDate.of(2100, 12, 31).atStartOfDay();
-        Slice<FeedResponse> recent10Feeds = feedService.getFeedsOfMemberId(writer.getId(), maxDateTime);
-
-        List<String> feedTexts = recent10Feeds.stream().map(FeedResponse::textContent).toList();
-        assertThat(recent10Feeds.hasNext()).isTrue();
-        assertThat(feedTexts).containsExactly(
-                "feed19", "feed18", "feed17", "feed16", "feed15",
-                "feed14", "feed13", "feed12", "feed11", "feed10"
-        );
-    }
-
-    private List<Feed> create20FeedsWith(Tag tag) {
-        Member member = Member.builder()
-                .username("writer")
-                .encryptedPassword("dqwdasda")
-                .email("d1wd@wdwfef.com")
-                .build();
-        Member writer = memberRepository.save(member);
-
-        List<Feed> feeds = new ArrayList<>();
-
-        List<Tag> tagList = List.of(tag);
-        for (int i = 0; i < 20; i++) {
-            Feed feed = Feed.builder()
-                    .member(writer)
-                    .textContent("#" + i)
-                    .tags(tagList)
-                    .longitude(11.11)
-                    .latitude(11.11)
-                    .build();
-            feeds.add(feed);
-        }
-        return feeds;
-    }
 }
